add first principle reasoning, update footer and favicon
