+++ index.html
@@ -2,9 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/thinker.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Canvasly — Business Model Canvas</title>
+    <title>Canvasly — Business Model Canvas & First Principle Reasoning</title>
     <style>
       * { margin: 0; padding: 0; box-sizing: border-box; }
       html, body { background: #0f0f0f; }
